Amendment to previous commit
The last commit changed the single-site capability to 'edit_files'.
It is planned that AceIDE will obey the "DISALLOW_FILE_EDIT"
constant, with an ability to disobey the constant via a filter.
This constant removes the edit_files capability, forcing us to
obey it. Let's use 'create_users' instead.

This also adds multiple permissions checks throughout the GitOps
module.