Changed AceIDE startup check wording
AceIDE provides the option to modify the root directory. Some of the
wording in the startup check is hard-coded to refer to wp-content in
place of the root directory. By default, this behaviour is not
noticable, however on changing the root directory, this information
becomes wrong. Changed wording to 'root' directory.

TODO: The startup checks two more directories - plugins and themes.
The wording remains the same. Not sure whether to remove these checks
or ensure they always go to wp-content unless wp-content is not
a part of the root's accessible tree.