Stop assumptions as to the root directory being made during startup checks
Thus far, AceIDE has had the ability to change the root directory. However,
startup checks have continued to assume that the root directory is set as
wp-content. This is usually fine, as it is the default value. Upon changing
this value, it makes the startup checker give us wrong information.

The startup checker will now always check the correct plugins and themes
directories as long as the wp-content is accessible from the root
directory, and in the case it is not it will output the correct message
accordingly.