Added FileOps::create_new - fixing commit 09f353a
Commit 09f353a introduced some modularisation into the
codebase in an attempt to better conform to SRP, or at
least to provide some separation of concerns between
classes in AceIDE. However during the migration of
file operations into the FileOps class, the ajax handler
for creating new files and folders was deleted. This
has broken the create file/folder feature.

We simply add the handler back in, to the FileOps class
to bring this feature back to working order.