Re-organized find and settings dialog code.
Upon removing the version of jQuery UI that shipped with AceIDE,
a number of other issues arose with dialogs (ie. find, settings,
git). This fixes the Z order as well as the close button being
mislocated, and places the HTML inside the wpbody-content so it
may sit in the Z order correctly.